src/test/java/se/kth/ControllerTest.java
package se.kth.controller;

import static org.mockito.Mockito.*;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import se.kth.integration.*;

public class ControllerTest {

    private Printer mockPrinter;
    private SalesLog mockSalesLog;
    private Controller controller;

    @BeforeEach
    public void setUp() {
        // Initialize the mock objects
        mockPrinter = mock(Printer.class);
        mockSalesLog = mock(SalesLog.class);
        
        // Create a Controller instance with mock dependencies
        controller = new Controller(mockPrinter, mockSalesLog);
    }

    @Test
    public void constructor_shouldInjectDependencies() {
        // Test that the constructor correctly sets the dependencies
        // Since this is private internal state, typically we'd use this test to ensure no exceptions, etc.
        // Here we demonstrate a simple instantiation check
        assertDoesNotThrow(() -> new Controller(mockPrinter, mockSalesLog));
    }

    @Test
    public void createAccountSystem_shouldInstantiateAccountingSystem() {
        // This will verify that an AccountingSystem is created when createAccountSystem is called
        controller.CreateAccountSystem();
        // Verify that a new AccountingSystem has been instantiated
        // Since we can't check internal instantiation directly without a reference, this might be a point to consider design changes
        // For this example, assume there's some internal logic or a static factory that we can't mock without a concrete reference
        assertTrue(true); // Placeholder for demonstration
    }

    @Test
    public void createInventorySystem_shouldInstantiateInventorySystem() {
        // This will verify that an InventorySystem is created when createInventorySystem is called
        controller.CreateInventorySystem();
        // Verify that a new InventorySystem has been instantiated
        assertTrue(true); // Similar placeholder due to the same reasons as above
    }

    @Test
    public void createDiscountHandler_shouldInstantiateDiscountHandler() {
        // This will verify that a DiscountHandler is created when createDiscountHandler is called
        controller.CreateDiscountHandler();
        // Verify that a new DiscountHandler has been instantiated
        assertTrue(true); // Again, a placeholder
    }
}

